requirements.txt
streamlit
replicate
pillow